Update README.md
A visionary artist and humanitarian ,. Rasta Mesu now creates the Mayan-Omniverse within this blockchain are tools for human transformation and metamorphosis. In orde for humans to transform and teleport into other dimentions the inernet must come of age. The Vision Quest in the Mayan Omniverse makes this possible. " There are more atoms in the human eye than stars in the galaxy. By the way, check out my awesome work.